etl/requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
python-dateutil  # для relativedelta